[nrf noup] dts: arm: nordic: Use PSA RNG driver
squash! [nrf noup] dts: choose cryptocell for entropy when available

This makes the Nordic devices nRF9160 and nRF5340
compatible with the new PSA RNG driver which
exists in Zephyr.
This is a noup commit because the CryptoCell is
only available downstream so the changes that
we do here didn't exist in Zephyr anyway.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>