[nrf fromlist] drivers: entropy: PSA crypto RNG driver default
Enable the PSA RNG driver by default. This option
will only be enabled when BUILD_WITH_TFM is enabled
and a device with the required compatible field
(zephyr,psa-crypto-rng) is defined in the device tree.
When a vendor includes such a device and enables the
ENTROPY_GENERATOR subsystem it is fair to assume
that wants to use the PSA Crypto RNG driver.

Upstream PR:
https://github.com/zephyrproject-rtos/zephyr/pull/52299

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>